ci(coverage): split wasm-bindgen wrappers and ignore in codecov
Move `start` and `evaluate` into a dedicated `bindings.rs` so the
`#[wasm_bindgen]` entry points (whose bodies are intrinsically
unreachable from native `cargo test`) can be excluded from patch
coverage without losing signal on the testable `run` / `PlaygroundBridge`
surface in `lib.rs`. Resolves the codecov/patch failure on PR #433.
